backend/requirements.txt
# ============================================================================
# Dependências de Produção - Flask Backend
# ============================================================================

# Framework Web
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Banco de Dados e ORM
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# Serialização e Validação
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Autenticação e Segurança
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
PyJWT==2.8.0

# Validação de Formulários
Flask-WTF==1.1.1
WTForms==3.0.1

# Geração de Documentos
reportlab==4.0.4
python-docx==0.8.11
Pillow==10.0.1
weasyprint==60.1

# Gráficos e Imagens
matplotlib==3.7.2
plotly==5.16.1

# Utilitários
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0
python-dateutil==2.8.2

# Cache e Performance
Redis==5.0.1
Flask-Caching==2.1.0

# Logs e Monitoramento
sentry-sdk[flask]==1.32.0

# Servidor de Produção
gunicorn==21.2.0

# Validação de CPF/CNPJ
python-cpf==1.1.0

# Manipulação de Timezone
pytz==2023.3

# Excel export
openpyxl==3.1.2
xlsxwriter==3.1.7

# Email
Flask-Mail==0.9.1

# Processamento de Imagens para Roda da Vida
opencv-python-headless==4.8.1.78
